HaikuDepot: Remove Custom List
Closes #15534

Reviewed-on: https://review.haiku-os.org/c/haiku/+/3758

+++ src/apps/haikudepot/ui/PackageInfoView.cpp
@@ -591,9 +591,9 @@ private:
 			return true;
 		if (actions.size() != fPackageActions.size())
 			return true;
-		if (fButtons.CountItems() != actions.size())
+		if (fButtons.CountItems() != static_cast<int32>(actions.size()))
 			return true;
-		for (int i = 0; (i < actions.size()); i++) {
+		for (int i = 0; i < static_cast<int32>(actions.size()); i++) {
 			if (actions[i]->Type() != fPackageActions[i]->Type())
 				return true;
 		}
